+++ src/css/base.scss
@@ -19,7 +19,7 @@ body {
   @include flex-column();
 }
 
-/* Image Section */
+/* IMAGE SECTION */
 
 img {
   @include sizing(100%, 99%);
@@ -35,7 +35,7 @@ img {
   @include margin-auto();
 }
 
-/* Navigation Bar Section */
+/* NAVIGATION BAR SECTION */
 
 h1 {
   font-size: 80px;
@@ -73,7 +73,7 @@ h1 {
   border: 5px solid #DAA520;
 }
 
-/* The Modal (background) */
+/* MODAL STYLING */
 
 .modal {
   display: none;
@@ -130,7 +130,7 @@ h1 {
   }
 }
 
-/* User Dashboard Section */
+/* USER DASHBOARD SECTION */
 
 h4 {
   font-size: $font-size24;
@@ -293,7 +293,7 @@ grid-template-areas: "aside left"
   padding: 30px;
 }
 
-/* Login Section */
+/* LOGIN SECTION */
 
 .login-section {
   @include flex-column();
@@ -319,7 +319,7 @@ grid-template-areas: "aside left"
   margin: 0px 10px 0px 0px;
 }
 
-/* Footer Section */
+/* FOOTER SECTION */
 
 .footer-container {
   text-align: center;
