+++ src/domUpdates.js
@@ -1,31 +1,41 @@
 import moment from 'moment';
 moment().format();
+import {
+  bookingsApi,
+  postBooking
+} from './apiCalls';
+import {currentUser, roomsData, usersData, bookingsData, verifyCustomerLogin, getAvailableRooms, getBookingData} from './scripts';
+
 
 /* QUERY SELECTORS */
-const loginButton = document.getElementById('loginButton')
+
+const loginButton = document.getElementById('loginButton');
+const dashboardButton = document.getElementById('userDashboardButton');
 const loginModal = document.getElementById('loginModal')
 const loginMessage = document.getElementById('loginMessage')
-const dashboardButton = document.getElementById('userDashboardButton');
 const welcomeMsg = document.getElementById('welcomeMsg');
 const totalSpent = document.getElementById('totalCost');
-
-
+const searchRoomsButton = document.getElementById('findRoomButton');
+const clearSearchButton = document.getElementById('clearSearchButton');
+const loginSubmitButton = document.getElementById('submitLoginButton');
+const modal = document.getElementById("myModal");
+const span = document.getElementsByClassName("close")[0];
+const spanLogin = document.getElementById("loginClose");
 const reservationInfo = document.getElementById("reservationInfo")
 const upcomingReservations = document.getElementById("upcoming");
 const pastReservations = document.getElementById("past")
-
 const availableRoomsSection = document.getElementById("availableRooms")
 const searchDate = document.getElementById("bookingDate");
 const searchRoomType = document.getElementById("roomType");
 const usernameInput = document.getElementById("username");
 const passwordInput = document.getElementById("password");
 
-function populateBookings(bookings, roomsData, reservationSection) {
+/* FUNCTIONS */
 
+function populateBookings(bookings, roomsData, reservationSection) {
   reservationSection.innerHTML = '';
 	
   bookings.forEach(booking => {
-
     const bookedRoom = roomsData.find(room => {
       return booking.roomNumber === room.number;
     });
@@ -51,8 +61,8 @@ function populateBookings(bookings, roomsData, reservationSection) {
 }
 
 let domUpdates = {
+
   renderDashboard(customer, roomsData) {
-    console.log(customer)
     welcomeMsg.innerText = customer.greetCurrentUser();
     totalSpent.innerText = customer.calculateTotalSpent(roomsData);
 
@@ -63,25 +73,23 @@ let domUpdates = {
     if (customer.upcomingBookings.length > 0) {
       populateBookings(customer.upcomingBookings, roomsData, upcomingReservations)
     } else {
-      upcomingReservations.innerHTML += `<p class="reservation-msg">Sorry, you have no upcoming reservations. Book one now!</p>`;
+      upcomingReservations.innerHTML += `<p class="reservation-msg">Sorry, you have no upcoming reservations. Book now!</p>`;
     }
 
     if (customer.pastBookings.length > 0) {
       populateBookings(customer.pastBookings, roomsData, pastReservations)
     } else {
-      pastReservations.innerHTML += `<p class="reservation-msg">Sorry, you have no upcoming reservations. Book one now!</p>`;
+      pastReservations.innerHTML += `<p class="reservation-msg">Sorry, you have no upcoming reservations. Book now!</p>`;
     }
   },
+
   renderAvailableRooms(searchCriteria, currentUser) {
-		
     availableRoomsSection.innerHTML = '';
-    console.log(moment(new Date(searchCriteria.date)).unix())
-    console.log(searchCriteria.date)
-    console.log(moment(new Date()).unix())
-    const formattedDate = moment(new Date()).format("YYYY-MM-DD");
+    const formattedDate = moment(new Date()).format('YYYY-MM-DD');
     if (moment(new Date(searchCriteria.date)).unix() < moment(new Date(formattedDate)).unix()) {
       availableRoomsSection.innerHTML = `
-				<p>Please select today or a date in the future
+				<p>
+        Please select today or a date in the future
 				</p>
 				`;
     } else if (currentUser.availableRooms.length < 1) {
@@ -93,7 +101,6 @@ let domUpdates = {
 			</p>
 			`;
     }
-
     currentUser.availableRooms.forEach(room => {
       availableRoomsSection.innerHTML += `
         <div class="room-card" id="${room.number}">
@@ -124,52 +131,116 @@ let domUpdates = {
           `;
     });
   },
+
   getSearchCriteria() {
     const date = searchDate.value;
     const type = searchRoomType.value;
 
-    return {date, type}
+    return {date, type};
   },
+  
   clearSearchCriteria() {
     searchDate.value = '';
     searchRoomType.selectedIndex = "0";
     availableRoomsSection.innerHTML = '';
   },
+
   clearRoomSearch() {
     setTimeout(() => {
       domUpdates.clearSearchCriteria();
     }, 2000);
   },
+
   getLoginInfo() {
-    let username = usernameInput.value
-    let password = passwordInput.value
-    console.log(username)
-    console.log(password)
-    return { username, password }
+    let username = usernameInput.value;
+    let password = passwordInput.value;
+    return { username, password };
   },
+
   displaySuccessfulLogin() {
     setTimeout(() => {
-      console.log("should print after 2 secs")
       dashboardButton.classList.remove('hidden');
       loginButton.innerText = 'Log Out';
       domUpdates.closeModal(loginModal);
-      loginMessage.innerText = "";
-      usernameInput.value = "";
-      passwordInput.value = "";
+      loginMessage.innerText = '';
+      usernameInput.value = '';
+      passwordInput.value = '';
     }, 2000);
-    loginMessage.innerText = "You have successful logged in!";
+    loginMessage.innerText = 'You have successfully logged in!';
   },
+
   showModal(modal) {
     modal.classList.add("showModal")
     modal.classList.remove("hideModal")
   },
+
   closeModal(modal) {
     modal.classList.add("hideModal")
     modal.classList.remove("showModal")
   },
+
   showBookingStatus(roomInfoDiv, info) {
     roomInfoDiv.innerHTML += `<p>${info.message}</p>`
   }
 }
+
+
+/* EVENT LISTENERS */
+
+dashboardButton.addEventListener('click', () => {
+  domUpdates.renderDashboard(currentUser, roomsData);   
+  domUpdates.showModal(modal);       
+})
+
+searchRoomsButton.addEventListener('click', () => {
+  const searchCriteria = domUpdates.getSearchCriteria();
+  const availableRooms = getAvailableRooms(searchCriteria);
+  currentUser.setAvailableRooms(availableRooms);
+  domUpdates.renderAvailableRooms(searchCriteria, currentUser);
+
+  // Adding Event listeners to each card rendered from the rooms search
+
+  const bookings = document.querySelectorAll(".room-card");
+  bookings.forEach(booking => {
+    const button = booking.querySelector(".book-room-button");
+    const roomInfo = booking.querySelector(".room-info")
+    button.addEventListener('click', function (event) {
+      const bookingData = getBookingData(event.target.value);
+      postBooking(bookingData, currentUser.id)
+        .then((message) => {
+          domUpdates.showBookingStatus(roomInfo, message);
+          bookingsApi().then(data => {
+            currentUser.getBookings(data.bookings, new Date())
+            domUpdates.renderDashboard(currentUser, roomsData)
+          });
+        })
+        .then(domUpdates.clearRoomSearch())
+        .catch(err => domUpdates.showBookingStatus(roomInfo, err));
+    })
+  })
+})
+
+clearSearchButton.addEventListener('click', () => {
+  domUpdates.clearSearchCriteria();
+})
+
+loginSubmitButton.addEventListener('click', () => {
+  const loginData = domUpdates.getLoginInfo();
+  verifyCustomerLogin(loginData);
+})
+
+span.addEventListener('click', () => {
+  domUpdates.closeModal(modal);
+})
+
+spanLogin.addEventListener('click', () => {
+  domUpdates.closeModal(loginModal);
+})
+
+window.addEventListener('click', (event) => {
+  if (event.target === modal || event.target === loginModal) {
+    domUpdates.closeModal(event.target);
+  }
+});
 
 export default domUpdates;
